+++ trixnity-messenger-compose-view/src/androidMain/kotlin/de/connect2x/messenger/compose/view/settings/NotificationsSettings.android.kt
@@ -25,27 +25,33 @@ private fun pushChannelId(userId: UserId, config: MatrixMessengerConfiguration) 
 }"
 
 @Composable
-fun DeviceSettingsButton(viewModel: NotificationSettingsSingleAccountViewModel) {
+fun DeviceSettingsButton(viewModel: NotificationSettingsSingleAccountViewModel, enabled: Boolean) {
     val context = LocalContext.current
     val messengerConfig = DI.get<MatrixMessengerConfiguration>()
     val i18n = DI.get<I18nView>()
-    Button(onClick = {
-        val intent: Intent = Intent(android.provider.Settings.ACTION_CHANNEL_NOTIFICATION_SETTINGS).putExtra(
-            android.provider.Settings.EXTRA_APP_PACKAGE,
-            context.packageName
-        ).putExtra(
-            android.provider.Settings.EXTRA_CHANNEL_ID, pushChannelId(viewModel.account, messengerConfig)
-        )
-        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
-        context.startActivity(intent)
-    }) {
+    Button(
+        onClick = {
+            val intent: Intent = Intent(android.provider.Settings.ACTION_CHANNEL_NOTIFICATION_SETTINGS).putExtra(
+                android.provider.Settings.EXTRA_APP_PACKAGE,
+                context.packageName
+            ).putExtra(
+                android.provider.Settings.EXTRA_CHANNEL_ID, pushChannelId(viewModel.account, messengerConfig)
+            )
+            intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+            context.startActivity(intent)
+        },
+        enabled = enabled
+    ) {
         Text(i18n.notificationsSettingsPlatform())
     }
 }
 
 @SuppressLint("UnrememberedMutableState")
 @Composable
-internal actual fun ColumnScope.PlatformNotificationSettings(viewModel: NotificationSettingsSingleAccountViewModel) {
+internal actual fun ColumnScope.PlatformNotificationSettings(
+    viewModel: NotificationSettingsSingleAccountViewModel,
+    enabled: Boolean
+) {
     val i18n = DI.get<I18nView>()
     val permissionNecessary = viewModel.notificationPermissionsNecessary.collectAsState().value
 
@@ -57,5 +63,5 @@ internal actual fun ColumnScope.PlatformNotificationSettings(viewModel: Notifica
         }
         SmallSpacer()
     }
-    DeviceSettingsButton(viewModel)
+    DeviceSettingsButton(viewModel, enabled)
 }